Move convert_basename_to_relative_paths and get_relative_paths to io.py
